Add support for deleting tokens
Add a trash icon that deletes the given token.  This doesn't
refresh the table and the UI for deleting the current session is
awful, but it's a start.

As part of this work, refactor into components and add context
state to store the results of login, which will eventually be used
for errors as well.
