+++ ui/package.json
@@ -17,6 +17,7 @@
     "gatsby": "^2.26.1",
     "react": "^16.12.0",
     "react-dom": "^16.12.0",
+    "react-icons": "^4.1.0",
     "react-table": "^7.6.2"
   },
   "devDependencies": {
